added input validation

+++ GSSS-Backend/src/main/java/ca/mcgill/ecse321/GSSS/service/BusinessHourService.java
@@ -18,25 +18,6 @@ public class BusinessHourService {
   BusinessHourRepository businessHourRepository;
 
   /**
-   * method that creates new business hour
-   * 
-   * @author Habib Jarweh
-   * @param weekday desired day of the week
-   * @param startTime desired start time of businesshour
-   * @param endTime desired end time of businesshour
-   * @return business hour we created
-   */
-  @Transactional
-  public BusinessHour createBusinessHour(Weekday weekday, Time startTime, Time endTime) {
-    BusinessHour businessHour = new BusinessHour();
-    businessHour.setWeekday(weekday);
-    businessHour.setStartTime(startTime);
-    businessHour.setEndTime(endTime);
-    businessHourRepository.save(businessHour);
-    return businessHour;
-  }
-
-  /**
    * method that gets business hour by weekday
    * 
    * @author Habib Jarweh
@@ -45,6 +26,9 @@ public class BusinessHourService {
    */
   @Transactional
   public BusinessHour getBusinessHourByWeekday(Weekday weekday) {
+    if (weekday == null) {
+      throw new IllegalArgumentException("weekday of business hour cannot be null! ");
+    }
     return businessHourRepository.findBusinessHourByWeekday(weekday);
   }
 
@@ -60,6 +44,44 @@ public class BusinessHourService {
   }
 
   /**
+   * method that creates new business hour
+   * 
+   * @author Habib Jarweh
+   * @param weekday desired day of the week
+   * @param startTime desired start time of businesshour
+   * @param endTime desired end time of businesshour
+   * @return business hour we created
+   */
+  @Transactional
+  public BusinessHour createBusinessHour(Weekday weekday, Time startTime, Time endTime) {
+    // Input validation
+    String error = "";
+    if (weekday == null) {
+      error += "weekday of business hour cannot be null! ";
+    }
+    if (startTime == null) {
+      error += "business hour start time cannot be null! ";
+    }
+    if (endTime == null) {
+      error += "business hour end time cannot be null! ";
+    }
+    if (endTime != null && startTime != null && endTime.before(startTime)) {
+      error += "business hour end time cannot be before business hour start time!";
+    }
+    error = error.trim();
+    if (error.length() > 0) {
+      throw new IllegalArgumentException(error);
+    }
+
+    BusinessHour businessHour = new BusinessHour();
+    businessHour.setWeekday(weekday);
+    businessHour.setStartTime(startTime);
+    businessHour.setEndTime(endTime);
+    businessHourRepository.save(businessHour);
+    return businessHour;
+  }
+
+  /**
    * method that deletes business hour
    * 
    * @author Habib Jarweh
@@ -68,28 +90,31 @@ public class BusinessHourService {
    */
   @Transactional
   public BusinessHour deleteBusinessHour(Weekday weekday) {
+    if (weekday == null) {
+      throw new IllegalArgumentException("weekday of business hour cannot be null! ");
+    }
     BusinessHour businessHour = businessHourRepository.findBusinessHourByWeekday(weekday);
     businessHourRepository.delete(businessHour);
     return businessHour;
 
   }
+
   /**
    * method to edit/modify a business hour of a certain day
    *
    * @author Chris Hatoum
    * @param day specific weekday
    * @param startTime start time of the day we want
-   * @param endTime  end time of the day we want
+   * @param endTime end time of the day we want
    * @return Buisness hours ( opening and closing ) of the day we want to update
    */
   @Transactional
-  public BusinessHour modifyBusinessHour(Weekday day, Time startTime, Time endTime){
+  public BusinessHour modifyBusinessHour(Weekday day, Time startTime, Time endTime) {
 
-     BusinessHour businessHour = businessHourRepository.findBusinessHourByWeekday(day);
-     businessHour.setStartTime(startTime);
-     businessHour.setEndTime(endTime);
-     businessHourRepository.save(businessHour);
-     return businessHour;
-   }
- }
-
+    BusinessHour businessHour = businessHourRepository.findBusinessHourByWeekday(day);
+    businessHour.setStartTime(startTime);
+    businessHour.setEndTime(endTime);
+    businessHourRepository.save(businessHour);
+    return businessHour;
+  }
+}
